# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ДомПульс API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/houses/{house_id}/contacts": {
      "get": {
        "tags": [
          "resident"
        ],
        "summary": "Contacts",
        "operationId": "contacts_api_v1_houses__house_id__contacts_get",
        "security": [
          {
            "MAXLaunchData": []
          }
        ],
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "House Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactsOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": [
          "resident"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          }
        },
        "security": [
          {
            "MAXLaunchData": []
          }
        ]
      }
    },
...
    "/api/v1/issues/{issue_id}": {
      "get": {
        "tags": [
          "resident"
        ],
        "summary": "Detail",
        "operationId": "detail_api_v1_issues__issue_id__get",
        "security": [
          {
            "MAXLaunchData": []
          }
        ],
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/issues/{issue_id}/photos/{photo_id}": {
      "get": {
        "tags": [
          "resident"
        ],
        "summary": "Resident Photo",
        "operationId": "resident_photo_api_v1_issues__issue_id__photos__photo_id__get",
        "security": [
          {
            "MAXLaunchData": []
          }
        ],
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          },
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Photo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/staff/houses/{house_id}/contacts": {
      "get": {
        "tags": [
          "Staff"
        ],
        "summary": "Contacts",
        "operationId": "contacts_api_v1_staff_houses__house_id__contacts_get",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "House Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactsOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Staff"
        ],
        "summary": "Update Contacts",
        "operationId": "update_contacts_api_v1_staff_houses__house_id__contacts_post",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "House Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactsInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactsOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/auth/login": {
      "post": {
        "tags": [
          "Staff"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_staff_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/staff/issues": {
      "get": {
        "tags": [
          "Staff"
        ],
        "summary": "Issues",
        "operationId": "issues_api_v1_staff_issues_get",
        "parameters": [
          {
            "name": "house_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "House Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IssueStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IssuePriority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffIssuePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/issues/{issue_id}": {
      "get": {
        "tags": [
          "Staff"
        ],
        "summary": "Issue Detail",
        "operationId": "issue_detail_api_v1_staff_issues__issue_id__get",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/issues/{issue_id}/status": {
      "post": {
        "tags": [
          "Staff"
        ],
        "summary": "Update Status",
        "operationId": "update_status_api_v1_staff_issues__issue_id__status_post",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/issues/{issue_id}/messages": {
      "post": {
        "tags": [
          "Staff"
        ],
        "summary": "Send Message",
        "operationId": "send_message_api_v1_staff_issues__issue_id__messages_post",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/photos/{photo_id}": {
      "get": {
        "tags": [
          "Staff"
        ],
        "summary": "Photo",
        "operationId": "photo_api_v1_staff_photos__photo_id__get",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Photo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/registrations": {
      "get": {
        "tags": [
          "Staff"
        ],
        "summary": "Registrations",
        "operationId": "registrations_api_v1_staff_registrations_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/registrations/{application_id}/{decision}": {
      "post": {
        "tags": [
          "Staff"
        ],
        "summary": "Decide Application",
        "operationId": "decide_application_api_v1_staff_registrations__application_id___decision__post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          },
          {
            "name": "decision",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Decision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContactsInput": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContactInput"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ContactsInput"
      },
      "ContactsOutput": {
        "properties": {
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContactInput"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "company_name",
          "items"
        ],
        "title": "ContactsOutput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IssueDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "category": {
            "$ref": "#/components/schemas/IssueCategory"
          },
          "priority": {
            "$ref": "#/components/schemas/IssuePriority"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "house_id": {
            "type": "integer",
            "title": "House Id"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "entrance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entrance"
          },
          "floor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor"
          },
          "zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone"
          },
          "apartment": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apartment"
          },
          "photo_count": {
            "type": "integer",
            "title": "Photo Count"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/PhotoOutput"
            },
            "type": "array",
            "title": "Photos"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/EventOutput"
            },
            "type": "array",
            "title": "History"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageOutput"
            },
            "type": "array",
            "title": "Messages"
          },
          "notifications": {
            "items": {
              "$ref": "#/components/schemas/DeliveryOutput"
            },
            "type": "array",
            "title": "Notifications"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "category",
          "priority",
          "status",
          "created_at",
          "house_id",
          "address",
          "author",
          "entrance",
          "floor",
          "zone",
          "apartment",
          "photo_count",
          "photos",
          "history",
          "messages",
          "notifications"
        ],
        "title": "IssueDetail"
      },
...
      "IssuePriority": {
        "type": "integer",
        "enum": [
          1,
          2,
          3
        ],
        "title": "IssuePriority"
      },
      "IssueResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "house_id": {
            "type": "integer",
            "title": "House Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "category": {
            "$ref": "#/components/schemas/IssueCategory"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "mine": {
            "type": "boolean",
            "title": "Mine"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "place": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/IssueHistory"
            },
            "type": "array",
            "title": "History"
          },
          "photo_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Photo Ids"
          }
        },
        "type": "object",
        "required": [
          "id",
          "house_id",
          "title",
          "description",
          "category",
          "status",
          "priority",
          "created_at",
          "mine",
          "address",
          "place",
          "history"
        ],
        "title": "IssueResponse"
      },
...
      "IssueStatus": {
        "type": "string",
        "enum": [
          "new",
          "in_progress",
          "resolved"
        ],
        "title": "IssueStatus"
      },
...
      "LoginInput": {
        "properties": {
          "login": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Login"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "login",
          "password"
        ],
        "title": "LoginInput"
      },
...
      "Profile": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "apartments": {
            "items": {
              "$ref": "#/components/schemas/ApartmentResponse"
            },
            "type": "array",
            "title": "Apartments"
          },
          "houses": {
            "items": {
              "$ref": "#/components/schemas/HouseResponse"
            },
            "type": "array",
            "title": "Houses"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "apartments",
          "houses"
        ],
        "title": "Profile"
      },
...
      "StaffIssuePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IssueRow"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "counts"
        ],
        "title": "StaffIssuePage"
      },
      "StatusInput": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "expected_status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "expected_status",
          "request_id"
        ],
        "title": "StatusInput"
      },
...
    },
    "securitySchemes": {
      "MAXLaunchData": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}